% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_codons_to_aa.R
\name{dna_codons_to_aa}
\alias{dna_codons_to_aa}
\title{Translate codon to their corresponding amino acid}
\usage{
dna_codons_to_aa(codons)
}
\arguments{
\item{codons}{The codon to be translated}
}
\value{
Corresponding amino acid
}
\description{
Translate codon to their corresponding amino acid
}
\examples{
dna_codons_to_aa("CGA")

}
