% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna.R
\name{random_dna}
\alias{random_dna}
\title{Generate a random strand of DNA}
\usage{
random_dna(l)
}
\value{
A string
}
\description{
Given a number of nucleotides (l), it returns a string of nucleotide of said length.
}
